+++ gameengine/src/main/java/de/mirkosertic/gameengine/arcaderacer/TrackElement.java
@@ -5,11 +5,11 @@ import de.mirkosertic.gameengine.type.Color;
 public class TrackElement {
 
     public final Road road;
-    public final float height;
+    public final double height;
     public final Color terrainColor;
     public final Sprite[] sprites;
 
-    public TrackElement(Road aRoad, float aHeight, Color aTerrainColor, Sprite[] aSprites) {
+    public TrackElement(Road aRoad, double aHeight, Color aTerrainColor, Sprite[] aSprites) {
         road = aRoad;
         height = aHeight;
         terrainColor = aTerrainColor;
